fix 'offsetof' bug on 64bits Windows.

+++ list.h
@@ -172,7 +172,7 @@ static inline void list_splice_init(struct list_head *list,
  * @member:	the name of the list_struct within the struct.
  */
 #define list_entry(ptr, type, member) \
-	((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+	((type *)((char *)(ptr)-(size_t)(&((type *)0)->member)))
 
 /**
  * list_for_each	-	iterate over a list
@@ -310,7 +310,7 @@ static inline void slist_splice_init(struct slist_head *list,
 }
 
 #define slist_entry(ptr, type, member) \
-	((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+	((type *)((char *)(ptr)-(size_t)(&((type *)0)->member)))
 
 #define slist_for_each(pos, head) \
 	for (pos = (head)->first.next; pos; pos = pos->next)
